Proj_11_04_14
Changed all end div tags and header to include logout
Need to change closet.php to modify if there is no user image
Need to add echo statements

+++ Closet.php
@@ -43,73 +43,51 @@ if (!(isset($_SESSION['login']) && $_SESSION['login'] != '')) {
 </head>
 
 
-
-<script>
-var password = "<? echo $_GET['pass'];  ?>";
-	function updateStock(id){
-		var txtbox = document.getElementById("description" + pk);
-	var url = "editCloset.php?pass="+password+"&action=update&id="+id+"&description="+txtbox.value;
-	window.location= url;
-	}
-	
-	function deleteStock(id){
-		var url="editCloset.php?pass="+password+"&action=delete+&id="+id;
-		window.location=url;
-	}	
-
-
-</script>
-
 <body>
 <!--Closet Profile --------------------------------- -->
-
 <div data-role="page" id="Closet">
-  <div data-role="header" data-position ="fixed">
-  <a rel="external" href="FastFash.php" data-icon="home" data-iconpos="left" data-direction="reverse" class="ui-btn-left"
-  		data-transition = "flip">Home</a>
-  <a href="#" class="ui-btn ui-btn-icon-left ui-icon-search ui-corner-all ui-shadow">Search</a>
-    <h1>Welcome to your Closet Profile!</h1>
-  </div>
+  	<div data-role="header" data-position ="fixed">
+  		<a rel="external" href="FastFash.php" data-icon="home" data-iconpos="left" data-direction="reverse" class="ui-btn-left" data-transition = "flip">Home</a>
+  		<a href ="page2.php" rel="external" data-role="button" data-icon="minus">Log out</a>
+		<h1>Closet Profile!</h1>
+	</div>
   
-    <!--Navigation Bar ----------------------------------->
-    <div data-role="navbar">
+<!--Navigation Bar ----------------------------------->
+<div data-role="navbar">
     	<ul>
     		<li><a rel="external" href="trends.php?pass=hello">Trends</a>
     		<li><a rel="external" href="FastFash.php#Outfit">Outfit of the Day</a>
     	</ul>
-    </div>
+</div>
     
-    <!--Closet profile content -----------------------------------> 
-  <div data-role="content" >
+<!--Closet profile content -----------------------------------> 
+<div data-role="content" >
   <div data-role="collapsible">
-      <h1>Shoes!</h1>
+      	<h1>Shoes!</h1>
       	<p>
       This is page four and should contain all the clothes owned by the user</p>
-      		<div data-role="collapsible">
+      	<div data-role="collapsible">
       		<h4>Heels</h4>
     			<ul data-role="listview" data-filter="true" data-filter-placeholder="Search closet..." data-inset="true">
     				<li><a href="#">Green</a></li>
     				<li><a href="#">Black</a></li>
-    				
-				</ul>   	
-	
-      		</div>
+    			</ul>   	
+		</div>
     	<div data-role="collapsible">
       		<h1>Pumps</h1>
-      	 	
-      		<p>This should contain a image and description of users pumps </p>
+      	 	<p>This should contain a image and description of users pumps </p>
       	</div>
    </div>
    
 
    
     <ul data-role="listview" data-filter="true" data-filter-placeholder="Search closet..." data-inset="true">
-    <li><a href="#">Shoes</a></li>
-    <li><a href="#">Tops</a></li>
-    <li><a href="#">Jeans</a></li>
-    <li><a href="#">Accessories</a></li>
-    <li><a href="#">Dresses</a></li>
-    <li><a href="#">Coats/Jackets</a></li>
+    	<li><a href="#">Shoes</a></li>
+    	<li><a href="#">Tops</a></li>
+    	<li><a href="#">Jeans</a></li>
+    	<li><a href="#">Accessories</a></li>
+    	<li><a href="#">Dresses</a></li>
+    	<li><a href="#">Coats/Jackets</a></li>
 	</ul>   	
 	
   	
@@ -137,7 +115,6 @@ if (mysqli_connect_errno()) {
 
 
 
-
 /* Select queries return a resultset, selects customer first names */
 if ($result = mysqli_query($link, "SELECT messages FROM comments")) {
     printf("\nThere has been %d messages submitted. \n", mysqli_num_rows($result));
@@ -152,8 +129,6 @@ if ($result = mysqli_query($link, "SELECT messages FROM comments")) {
  echo "This is for session variables";
 
 
-
-
 /* Selects the customers message from the database and prints to the screen */
 $mysqli = new mysqli("danu6.it.nuigalway.ie", "mydb1396cs", "da1sus", "mydb1396");
 if ($mysqli->connect_errno) {
@@ -162,27 +137,41 @@ if ($mysqli->connect_errno) {
 
 $mysqli->real_query("SELECT id, pass, username,name,image,description FROM closet WHERE username='{$_SESSION['username']}'");
 $res = $mysqli->use_result();
-echo '<form action="editCloset.php" method="post">';
-echo "The items currently in your closet are:...";
-echo "<table border=0><tr><td>Name</td><td>Image</td><td>Description</td><td>Delete</td><td>Edit</td><td>Add</td></tr>";
-while ($row = $res->fetch_assoc()) {
-    echo "<tr><td>{$row['name']}</td>";
-    echo "<td><img class='thumbnail' src={$row['image']} width='200' height='150' /></td>" ;
-    echo "<td>{$row['description']}</td>" ;
-	//echo "<td><input type=button value='delete' onClick='deleteStock({$row['id']});'> </td>";
-	echo "<td><a rel='external' href='editCloset.php?pass=hello&action=delete&id={$row['id']}' data-role='button' data-icon='delete'>Delete</a></td>";
-	echo "<td><a rel='external' href='editCloset.php?pass=hello&action=edit&id={$row['id']}' data-role='button' data-icon='plus'>Edit</a></td>";
-  	echo "<td><a rel='external' href='addNewCloset.php' data-role='button' data-icon='plus' data-theme='b'>Add</a></td></tr>";
+$row = $res->fetch_assoc();
+$image = "{$row['image']}";
 
-	//echo "<td><input type=button value='update' onClick='updateStock({$row['id']});'> </td></tr>";
+/*Check if there is already an image uploaded by the current user, if not send them a message to upload images to their profile*/
+	if ($image == null) {
+   				 echo "There are currently no items in your closet, please upload some images";
+   				 echo '<form action="editCloset.php" method="post">';
+				 echo "Add items to your closet:...";
+				 echo "<table border=0><tr><td></td></tr>";
+					while ($row = $res->fetch_assoc()) {
+    					echo "<td><a rel='external' href='addNewCloset.php' data-role='button' data-icon='plus' data-theme='b'>Add</a></td></tr>";
+    					}
+				echo "</table>";
 
+				}
+				
+/*Else the user already has  a profile and can now view their items*/				
+		else{
+		echo '<form action="editCloset.php" method="post">';
+		echo "The items currently in your closet are:...";
+		echo "<table border=0><tr><td>Name</td><td>Image</td><td>Description</td><td>Delete</td><td>Edit</td><td>Add</td></tr>";
+			while ($row = $res->fetch_assoc()) {
+    			echo "<tr><td>{$row['name']}</td>";
+    			echo "<td><img class='thumbnail' src={$row['image']} width='200' height='150' /></td>" ;
+    			echo "<td>{$row['description']}</td>" ;
+				echo "<td><a rel='external' href='editCloset.php?pass=hello&action=delete&id={$row['id']}' data-role='button' data-icon='delete'>Delete</a></td>";
+				echo "<td><a rel='external' href='editCloset.php?pass=hello&action=edit&id={$row['id']}' data-role='button' data-icon='plus'>Edit</a></td>";
+  				echo "<td><a rel='external' href='addNewCloset.php' data-role='button' data-icon='plus' data-theme='b'>Add</a></td></tr>";
     
-}
-echo "</table>";
+				}
+		echo "</table>";
 
+		}
 
-
-/* close connection */
+/* close connection to database*/
 
 mysqli_close($link);
 
@@ -191,20 +180,16 @@ mysqli_close($link);
 
 ?>
 
-
- <!--Footer -----------------------------------> 
-  <div data-role="footer" data-position ="fixed">
+<!--Footer -----------------------------------> 
+<div data-role="footer" data-position ="fixed">
    <a href="https://www.facebook.com/siobhan.collins.777" data-role="button" data-icon="plus">Add Me On Facebook</a>
-   <a href ="page2.php" rel="external" data-role="button" data-icon="minus">Log out</a>
-  </div>
+</div>
   
-      
-
+    
 <!-- --------------------------------- -->
 <!-- end content --> 
 
 </div> 
-
 
 <!-- end page --> 
 </div>
